Encapsulate the poster banner component
